jwt token flow

+++ src/main/kotlin/com/example/project24/config/JwtProperties.kt
@@ -5,6 +5,6 @@ import org.springframework.boot.context.properties.ConfigurationProperties
 @ConfigurationProperties("jwt")
 data class JwtProperties(
     val key: String,
-    val accessTokenExpiration: Int,
-    val refreshTokenExpiration: Int,
+    val accessTokenExpiration: Long,
+    val refreshTokenExpiration: Long,
 )
